[LOG] Set the log level enum in the log namespace instead of the logger class [LOG] Changed the structure of the logging system (the logger module and the actual logger are now distinct) [FS] Added a compressed source (wip) [FS] Added a memory stream (wip) [FS] Added a write function for strings [RES] The media manager can now not use the default types [OGL] Change buffer's and shader's bases inl files to cpp Changed priv namespaces to anonymous Several other minor changes

+++ trunk/include/djah/fs/memory_stream.hpp
@@ -11,7 +11,7 @@ namespace djah { namespace fs {
 		: public stream
 	{
 	public:
-		memory_stream(void *buffer, u32 size);
+		memory_stream(void *buffer, size_t size);
 
 		size_t size();
 		bool eof();
@@ -22,7 +22,7 @@ namespace djah { namespace fs {
 		size_t writeImpl(const byte* buff, size_t size);
 
 		boost::scoped_array<byte>	buffer_;
-		u32							buffer_size_;
+		size_t						buffer_size_;
 		byte						*cursor_;
 	};
